+++ src/Rask.Core/Browser/RaskBrowserJsonContext.cs
@@ -15,5 +15,6 @@ namespace Rask.Core.Browser;
 [JsonSerializable(typeof(GeolocationPosition))]
 [JsonSerializable(typeof(NetworkReading))]
 [JsonSerializable(typeof(SpeechOptions))]
+[JsonSerializable(typeof(ScreenInfo))]
 [JsonSerializable(typeof(Dictionary<string, string>))]
 internal sealed partial class RaskBrowserJsonContext : JsonSerializerContext;