+++ src/Rask.Server/RaskEndpointExtensions.cs
@@ -198,6 +198,7 @@ public static class RaskEndpointExtensions
         services.AddScoped<INetworkInfo, NetworkInfo>();
         services.AddScoped<IMediaQuery, MediaQuery>();
         services.AddScoped<ISpeechSynthesis, SpeechSynthesis>();
+        services.AddScoped<IScreenInfo, ScreenInfoReader>();
         services.AddScoped<ICookies, Cookies>();
         services.AddScoped<IPermissions, Permissions>();
         services.AddScoped<IVibration, Vibration>();